Added metrics server to zmq frontend and cleaned up benchmark packages
