+++ model_composition/containerized_utils/setup.py
@@ -2,16 +2,19 @@ from setuptools import setup
 import os
 
 setup(
-    name='zmq_client',
+    name='containerized_utils',
     version='1.0.0',
     description='A ZMQ-based client for querying Clipper',
     maintainer='Corey Zumar',
     maintainer_email='czumar@berkeley.edu',
     url='http://clipper.ai',
     packages=[
-        "zmq_client"
+        "containerized_utils",
+        "containerized_utils.zmq_client",
+        "containerized_utils.driver_utils"
     ],
     install_requires=[
         'numpy',
-        'pyzmq'
+        'pyzmq',
+        'futures_then'
     ])
