+++ model_composition/single_proc_utils/setup.py
@@ -2,14 +2,14 @@ from setuptools import setup
 import os
 
 setup(
-	name='composition_utils',
+	name='single_proc_utils',
 	version='1.03',
 	description='utilities for benchmarking models used in Clipper model composition research',
 	maintainer='Corey Zumar',
 	maintainer_email='czumar@berkeley.edu',
 	url='http://clipper.ai',
 	packages=[
-		"composition_utils", "composition_utils.model", "composition_utils.driver"
+		"single_proc_utils", "single_proc_utils.model", "single_proc_utils.driver"
 	],
 	install_requires=[
 		'numpy',
