--- model_composition/zmq_client/zmq_client.egg-info/PKG-INFO
@@ -1,10 +1,0 @@
-Metadata-Version: 1.0
-Name: zmq-client
-Version: 1.0.0
-Summary: A ZMQ-based client for querying Clipper
-Home-page: http://clipper.ai
-Author: Corey Zumar
-Author-email: czumar@berkeley.edu
-License: UNKNOWN
-Description: UNKNOWN
-Platform: UNKNOWN
