--- model_composition/zmq_client/zmq_client.egg-info/SOURCES.txt
@@ -1,7 +1,0 @@
-zmq_client/__init__.py
-zmq_client/clipper_zmq_client.py
-zmq_client.egg-info/PKG-INFO
-zmq_client.egg-info/SOURCES.txt
-zmq_client.egg-info/dependency_links.txt
-zmq_client.egg-info/requires.txt
-zmq_client.egg-info/top_level.txt
