--- model_composition/zmq_client/zmq_client.egg-info/requires.txt
@@ -1,2 +1,0 @@
-numpy
-pyzmq
