--- model_composition/zmq_client/zmq_client.egg-info/top_level.txt
@@ -1,1 +1,0 @@
-zmq_client
